Add default myDisposeCallback
I was seeing console errors due to a null `myDisposeCallback`. No where
seems to be setting it, but `module.hot.apply` is still called. There
should probably be some type checks that the callback is callable, but
to keep the code clean I have provided a simple noop callback instead.